added CharHuh (also my formatter became stupid and i had to manually revert a bunch of stuff sorry)

+++ app/TypesWasm.hs
@@ -6,36 +6,35 @@ import Data.Bits
 import Control.Conditional (cond)
 import AST
 import Data.Char (ord, chr)
-import Data.Int (Int64)
 
 immShift :: Int
 immShift = 3
-immMask :: Int64
+immMask :: Int
 immMask = 0b111
-ptrMask :: Int64
+ptrMask :: Int
 ptrMask = 0b111
-typeBox :: Integer
+typeBox :: Int
 typeBox = 0b001
-typeCons :: Integer
+typeCons :: Int
 typeCons = 0b010
-typeVect :: Integer
+typeVect :: Int
 typeVect = 0b011
-typeStr :: Integer
+typeStr :: Int
 typeStr = 0b100
 intShift :: Int
 intShift = 1 + immShift
-maskInt :: Int64
+maskInt :: Int
 maskInt = 0b1111
 charShift :: Int
 charShift = 2 + immShift
-typeInt :: Int64
+typeInt :: Int
 typeInt = 0b0000
-typeChar :: Int64
+typeChar :: Int
 typeChar = 0b01000
-maskChar :: Int64
+maskChar :: Int
 maskChar = 0b11111
 
-valueToBits :: Expr -> Int64
+valueToBits :: Expr -> Int
 valueToBits e = case e of
     Int i -> shiftL (fromIntegral i) intShift
     Bool b -> (if b then 0b00111000 else 0b00011000)
@@ -44,7 +43,7 @@ valueToBits e = case e of
     Empty -> 0b10011000
     _ -> -1
 
-bitsToValue :: Int64 -> Expr
+bitsToValue :: Int -> Expr
 
 bitsToValue b = cond [(valueToBits (Bool True) == b, Bool True )
                   , (valueToBits (Bool False) == b, Bool False )
@@ -53,11 +52,11 @@ bitsToValue b = cond [(valueToBits (Bool True) == b, Bool True )
                   , (bitsIsChar b, Char (chr (shiftR (fromIntegral b) charShift)))
                   , (otherwise , Eof)]
 
-bitsIsInt :: Int64 -> Bool
+bitsIsInt :: Int -> Bool
 bitsIsInt b = typeInt == (b .&. maskInt)
 
-bitsIsChar :: Int64 -> Bool
+bitsIsChar :: Int -> Bool
 bitsIsChar b = typeChar == (b .&. maskChar)
 
-bitsIsImm :: Int64 -> Bool
+bitsIsImm :: Int -> Bool
 bitsIsImm b = 0 == (b .&. immMask)
